[HOMEPAGE] create route and all stuff to make homepage independent. Remove homepage from index by default and make it load with ajax automatically
