+++ ExpressApp1/public/stylesheets/css/homePage/homePageMap.css
@@ -29,7 +29,7 @@
 #google-map
 {
 	position: relative;
-	background-image: url('../../../images/img/homePage/bgMap1.jpg');
+	background-image: url('../../../images/img/homePage/bgmap1.jpg');
 	background-size: cover;
 }
