+++ ExpressApp1/views/index.ejs
@@ -41,7 +41,6 @@
 		<div class="block" id="mainPageBlock">
 
 			<div class="page shown" id="homePage">
-				<%- include('homePage/homePageView', {introBoxes: introBoxes, ourStoryBoxes: ourStoryBoxes, projectBoxes: projectBoxes}) %>
 			</div>
 			<div class="page" id="timelinePage">
 				<!--<h1>timeline Page</h1>-->
